# backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: maicivy API
  version: 1.0.0
  description: |
    Interactive CV platform with AI-powered cover letter generation.

    This API provides endpoints for:
    - Adaptive CV generation based on themes
    - AI-powered dual cover letter generation (motivation + anti-motivation)
    - Real-time analytics and visitor tracking
    - GitHub integration for automated CV enrichment
    - Professional timeline and profile detection

    ## Rate Limiting
    - AI letter generation: 5 requests/day with 2-minute cooldown between requests
    - Access gate: Requires 3 visits or detected recruiter/tech profile

    ## Authentication
    Session-based authentication using cookies (session_id).
  contact:
    name: Alexi
    email: contact@maicivy.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.maicivy.com
    description: Production server (if deployed)
...
paths:
...
  /api/v1/letters/generate:
    post:
      tags:
        - Letters
      summary: Generate cover letter pair
      description: |
        Generates a dual cover letter (motivation + anti-motivation) asynchronously.

        **Rate Limits:**
        - 5 generations per day
        - 2-minute cooldown between requests

        **Access Requirements:**
        - 3 visits to the site OR
        - Detected as recruiter/tech profile
      operationId: generateLetter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateLetterRequest'
            example:
              company_name: Google
              job_title: Senior Backend Engineer
              theme: backend
      responses:
        '202':
          description: Generation job queued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LetterGenerationResponse'
              example:
                job_id: 550e8400-e29b-41d4-a716-446655440000
                status: queued
                message: "Génération en cours. Encore 4 génération(s) disponible(s) aujourd'hui."
                rate_limit_remaining: 4
        '400':
          description: Invalid request or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Session required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to enqueue job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/letters/jobs/{jobId}:
    get:
      tags:
        - Letters
      summary: Get job status
      description: Retrieves the status of a letter generation job (for polling)
      operationId: getJobStatus
      parameters:
        - name: jobId
          in: path
          required: true
          description: UUID of the generation job
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LetterJobStatus'
              examples:
                queued:
                  value:
                    job_id: 550e8400-e29b-41d4-a716-446655440000
                    status: queued
                    progress: 0
                    estimated_time: 45
                processing:
                  value:
                    job_id: 550e8400-e29b-41d4-a716-446655440000
                    status: processing
                    progress: 50
                    estimated_time: 20
                completed:
                  value:
                    job_id: 550e8400-e29b-41d4-a716-446655440000
                    status: completed
                    progress: 100
                    letter_motivation_id: "1"
                    letter_anti_motivation_id: "2"
                failed:
                  value:
                    job_id: 550e8400-e29b-41d4-a716-446655440000
                    status: failed
                    progress: 0
                    error: AI service timeout
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/letters/pair:
    get:
      tags:
        - Letters
      summary: Get letter pair by company
      description: Retrieves both motivation and anti-motivation letters for a company
      operationId: getLetterPair
      parameters:
        - name: company
          in: query
          required: true
          description: Company name
          schema:
            type: string
            example: Google
      responses:
        '200':
          description: Letter pair
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LetterPairResponse'
        '400':
          description: Company name required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Session required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No letters found for this company
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/letters/access-status:
    get:
      tags:
        - Letters
      summary: Check AI access status
      description: Checks if the current session has access to AI features
      operationId: getAccessStatus
      responses:
        '200':
          description: Access status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessStatusResponse'
              examples:
                no_access:
                  value:
                    has_access: false
                    current_visits: 1
                    required_visits: 3
                    visits_remaining: 2
                    access_granted_by: ""
                    message: "Encore 2 visite(s) nécessaire(s) pour débloquer l'IA"
                access_by_visits:
                  value:
                    has_access: true
                    current_visits: 3
                    required_visits: 3
                    visits_remaining: 0
                    access_granted_by: visits
                    message: "Accès aux fonctionnalités IA accordé"
                access_by_profile:
                  value:
                    has_access: true
                    current_visits: 1
                    required_visits: 3
                    visits_remaining: 2
                    profile_detected: recruiter
                    access_granted_by: profile
                    message: "Accès aux fonctionnalités IA accordé"
        '401':
          description: Session required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/letters/rate-limit-status:
    get:
      tags:
        - Letters
      summary: Check rate limit status
      description: Checks the current AI generation rate limit status
      operationId: getRateLimitStatus
      responses:
        '200':
          description: Rate limit status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitStatusResponse'
              example:
                daily_limit: 5
                daily_used: 2
                daily_remaining: 3
                reset_at: "2025-12-14 00:00:00"
                cooldown_active: false
                cooldown_remaining: 0
        '401':
          description: Session required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ====================
  # ANALYTICS ENDPOINTS
  # ====================
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: string

    # CV
...
    GenerateLetterRequest:
      type: object
      required:
        - company_name
      properties:
        company_name:
          type: string
          minLength: 2
          maxLength: 200
          example: Google
        job_title:
          type: string
          minLength: 2
          maxLength: 200
          example: Senior Backend Engineer
        theme:
          type: string
          enum: [backend, frontend, fullstack, devops, data, ai]
          example: backend
    LetterGenerationResponse:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued]
        message:
          type: string
        rate_limit_remaining:
          type: integer
    LetterJobStatus:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, processing, completed, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        letter_motivation_id:
          type: string
          nullable: true
        letter_anti_motivation_id:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        estimated_time:
          type: integer
          description: Estimated remaining time in seconds
          nullable: true
    LetterDetailResponse:
      type: object
      properties:
        id:
          type: string
        company_name:
          type: string
        letter_type:
          type: string
          enum: [motivation, anti_motivation]
        content:
          type: string
        created_at:
          type: string
          format: date-time
        ai_model:
          type: string
          example: claude-3-5-sonnet-20241022
        tokens_used:
          type: integer
        generation_ms:
          type: integer
        cost:
          type: number
          format: float
          description: Estimated cost in USD
        pdf_url:
          type: string
          format: uri
    LetterPairResponse:
      type: object
      properties:
        motivation_letter:
          $ref: '#/components/schemas/LetterDetailResponse'
          nullable: true
        anti_motivation_letter:
          $ref: '#/components/schemas/LetterDetailResponse'
          nullable: true
        company_name:
          type: string
        company_info:
          type: object
          nullable: true
...
    AccessStatusResponse:
      type: object
      properties:
        has_access:
          type: boolean
        current_visits:
          type: integer
        required_visits:
          type: integer
        visits_remaining:
          type: integer
        profile_detected:
          type: string
        access_granted_by:
          type: string
          enum: [visits, profile, ""]
        message:
          type: string
    RateLimitStatusResponse:
      type: object
      properties:
        daily_limit:
          type: integer
        daily_used:
          type: integer
        daily_remaining:
          type: integer
        reset_at:
          type: string
          format: date-time
        cooldown_active:
          type: boolean
        cooldown_remaining:
          type: integer
          description: Cooldown remaining in seconds

    # Timeline
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: Session-based authentication using cookies
security:
  - sessionCookie: []